Revert "fixed day format"
This reverts commit 3d6e799b77f66a89766087f998b5385aea369e01.

+++ src/helpers.js
@@ -32,7 +32,7 @@ const mm = String(currentDate.getMonth() + 1).padStart(2, '0');
 const dd = String(currentDate.getDate()).padStart(2, '0');
 const yyyy = String(currentDate.getFullYear()).slice(-2);
 const formattedDate = `${mm}-${dd}-${yyyy}`;
-const day = currentDate.toLocaleDateString('en-US', { weekday: 'short' });
+const day = currentDate.toLocaleDateString('en-US', { weekday: 'long' });
 
 //create shift
